+++ lib/intcomputer/operation/opfuncs.go
@@ -1,13 +1,13 @@
 package operation
 
-var OpFuncs = map[int]OpFunc{
+var OpFuncs = map[int64]OpFunc{
 	// Opcode 1 adds together numbers read from two positions and stores the result in a third position. The three
 	//integers immediately after the opcode tell you these three positions - the first two indicate the positions from
 	//which you should read the input values, and the third indicates the position at which the output should be stored.
 	1: func(op Operation, c Computer) {
 		params := op.ParamValues(c, 2)
 		rel := false
-		if op.ParamMode(len(params)) == RelativeMode {
+		if op.ParamMode(int64(len(params))) == RelativeMode {
 			rel = true
 		}
 		c.WritePosition(c.NextInt(), rel, params[0]+params[1])
@@ -18,7 +18,7 @@ var OpFuncs = map[int]OpFunc{
 	2: func(op Operation, c Computer) {
 		params := op.ParamValues(c, 2)
 		rel := false
-		if op.ParamMode(len(params)) == RelativeMode {
+		if op.ParamMode(int64(len(params))) == RelativeMode {
 			rel = true
 		}
 		c.WritePosition(c.NextInt(), rel, params[0]*params[1])
@@ -63,12 +63,12 @@ var OpFuncs = map[int]OpFunc{
 	//the third parameter. Otherwise, it stores 0.
 	7: func(op Operation, c Computer) {
 		params := op.ParamValues(c, 2)
-		val := 0
+		val := int64(0)
 		if params[0] < params[1] {
 			val = 1
 		}
 		rel := false
-		if op.ParamMode(len(params)) == RelativeMode {
+		if op.ParamMode(int64(len(params))) == RelativeMode {
 			rel = true
 		}
 		c.WritePosition(c.NextInt(), rel, val)
@@ -78,12 +78,12 @@ var OpFuncs = map[int]OpFunc{
 	//third parameter. Otherwise, it stores 0.
 	8: func(op Operation, c Computer) {
 		params := op.ParamValues(c, 2)
-		val := 0
+		val := int64(0)
 		if params[0] == params[1] {
 			val = 1
 		}
 		rel := false
-		if op.ParamMode(len(params)) == RelativeMode {
+		if op.ParamMode(int64(len(params))) == RelativeMode {
 			rel = true
 		}
 		c.WritePosition(c.NextInt(), rel, val)
